rename AddWallet folder to CreateWallet
