+++ WalletWasabi.Fluent/ViewModels/AddWalletPageViewModel.cs
@@ -9,8 +9,8 @@ using WalletWasabi.Wallets;
 using WalletWasabi.Stores;
 using NBitcoin;
 using WalletWasabi.Fluent.ViewModels.Dialogs;
-using WalletWasabi.Fluent.ViewModels.AddWallet;
 using System.Threading.Tasks;
+using WalletWasabi.Fluent.ViewModels.CreateWallet;
 using WalletWasabi.Gui.Validation;
 using WalletWasabi.Models;
 using WalletWasabi.Userfacing;
