+++ WalletWasabi.Fluent/ViewModels/CreateWallet/ConfirmRecoveryWordsViewModel.cs
@@ -1,17 +1,16 @@
-using DynamicData;
-using DynamicData.Binding;
-using ReactiveUI;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
 using System.Reactive.Linq;
 using System.Windows.Input;
+using DynamicData;
+using DynamicData.Binding;
+using ReactiveUI;
 using WalletWasabi.Blockchain.Keys;
-using WalletWasabi.Gui.ViewModels;
 using WalletWasabi.Wallets;
 
-namespace WalletWasabi.Fluent.ViewModels.AddWallet
+namespace WalletWasabi.Fluent.ViewModels.CreateWallet
 {
 	public class ConfirmRecoveryWordsViewModel : RoutableViewModel
 	{
