+++ WalletWasabi.Fluent/Views/CreateWallet/RecoveryWordsView.axaml
@@ -7,7 +7,7 @@
              xmlns:behaviors="clr-namespace:WalletWasabi.Fluent.Behaviors;assembly=WalletWasabi.Fluent"
              xmlns:i="clr-namespace:Avalonia.Xaml.Interactivity;assembly=Avalonia.Xaml.Interactivity"
              d:DesignWidth="600" d:DesignHeight="500"
-             x:Class="WalletWasabi.Fluent.Views.AddWallet.RecoveryWordsView">
+             x:Class="WalletWasabi.Fluent.Views.CreateWallet.RecoveryWordsView">
   <c:ContentArea Title="Recovery words"
                  Caption="Write down the following numbered security words in order. These can be used to recover your wallet in the future. It is very important you keep them in a safe place."
                  EnableCancel="True" CancelContent="Cancel"
